+++ saleor/graphql/channel/types.py
@@ -16,6 +16,7 @@ from ...permission.enums import (
     ChannelPermissions,
     CheckoutPermissions,
     OrderPermissions,
+    PaymentPermissions,
 )
 from ..account.enums import CountryCodeEnum
 from ..core import ResolveInfo
@@ -28,12 +29,15 @@ from ..core.descriptions import (
     ADDED_IN_313,
     ADDED_IN_314,
     ADDED_IN_315,
+    ADDED_IN_316,
     DEPRECATED_IN_3X_FIELD,
+    DEPRECATED_PREVIEW_IN_316_FIELD,
     PREVIEW_FEATURE,
 )
 from ..core.doc_category import (
     DOC_CATEGORY_CHECKOUT,
     DOC_CATEGORY_ORDERS,
+    DOC_CATEGORY_PAYMENTS,
     DOC_CATEGORY_PRODUCTS,
 )
 from ..core.fields import PermissionsField
@@ -246,7 +250,11 @@ class OrderSettings(ObjectType):
         description=(
             "Determine the transaction flow strategy to be used. "
             "Include the selected option in the payload sent to the payment app, as a "
-            "requested action for the transaction." + ADDED_IN_313 + PREVIEW_FEATURE
+            "requested action for the transaction."
+            + ADDED_IN_313
+            + PREVIEW_FEATURE
+            + DEPRECATED_PREVIEW_IN_316_FIELD
+            + " Use `PaymentSettings.defaultTransactionFlowStrategy` instead."
         ),
     )
     delete_expired_orders_after = Day(
@@ -268,6 +276,21 @@ class OrderSettings(ObjectType):
     class Meta:
         description = "Represents the channel-specific order settings."
         doc_category = DOC_CATEGORY_ORDERS
+
+
+class PaymentSettings(ObjectType):
+    default_transaction_flow_strategy = TransactionFlowStrategyEnum(
+        required=True,
+        description=(
+            "Determine the transaction flow strategy to be used. "
+            "Include the selected option in the payload sent to the payment app, as a "
+            "requested action for the transaction." + ADDED_IN_316 + PREVIEW_FEATURE
+        ),
+    )
+
+    class Meta:
+        description = "Represents the channel-specific payment settings."
+        doc_category = DOC_CATEGORY_PAYMENTS
 
 
 class Channel(ModelObjectType):
@@ -373,6 +396,17 @@ class Channel(ModelObjectType):
         permissions=[
             ChannelPermissions.MANAGE_CHANNELS,
             CheckoutPermissions.MANAGE_CHECKOUTS,
+        ],
+    )
+    payment_settings = PermissionsField(
+        PaymentSettings,
+        description="Channel-specific payment settings."
+        + ADDED_IN_316
+        + PREVIEW_FEATURE,
+        required=True,
+        permissions=[
+            ChannelPermissions.MANAGE_CHANNELS,
+            PaymentPermissions.HANDLE_PAYMENTS,
         ],
     )
 
@@ -530,4 +564,10 @@ class Channel(ModelObjectType):
     def resolve_checkout_settings(root: models.Channel, _info):
         return CheckoutSettings(
             use_legacy_error_flow=root.use_legacy_error_flow_for_checkout
+        )
+
+    @staticmethod
+    def resolve_payment_settings(root: models.Channel, _info):
+        return PaymentSettings(
+            default_transaction_flow_strategy=root.default_transaction_flow_strategy,
         )